# lib/api-zod/src/generated/types/atsBreakdown.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CareerAI resume analysis and job matching API
 * OpenAPI spec version: 0.1.0
 */

export interface AtsBreakdown {
  skillsRelevance: number;
  keywords: number;
  experience: number;
  projects: number;
  education: number;
  completeness: number;
}