# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CareerAI resume analysis and job matching API
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  /** @nullable */
  name: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  location: string | null;
}