# lib/api-zod/src/generated/types/resumeDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CareerAI resume analysis and job matching API
 * OpenAPI spec version: 0.1.0
 */
import type { Resume } from './resume';
import type { ResumeAnalysis } from './resumeAnalysis';

export type ResumeDetail = Resume & ({
  analysis: ResumeAnalysis | null;
});