# lib/api-zod/src/generated/types/suggestion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CareerAI resume analysis and job matching API
 * OpenAPI spec version: 0.1.0
 */

export interface Suggestion {
  priority: number;
  title: string;
  detail: string;
}